+++ cdp/external_address.py
@@ -1,0 +1,5 @@
+from cdp.address import Address
+
+
+class ExternalAddress(Address):
+    """A class representing an external address."""